+++ src/pylibrelinkup/data_types.py
